# [bug#12009] Make ListBuffer's iterator fail-fast