--- a/src/library/scala/collection/mutable/ListBuffer.scala
+++ b/src/library/scala/collection/mutable/ListBuffer.scala
@@ -35,6 +35,7 @@ import scala.runtime.Statics.releaseFence
   *  @define mayNotTerminateInf
   *  @define willNotTerminateInf
   */
+@SerialVersionUID(-8428291952499836345L)
 class ListBuffer[A]
   extends AbstractBuffer[A]
      with SeqOps[A, ListBuffer, ListBuffer[A]]
@@ -42,6 +43,7 @@ class ListBuffer[A]
      with ReusableBuilder[A, immutable.List[A]]
      with IterableFactoryDefaults[A, ListBuffer]
      with DefaultSerializable {
+  @transient private[this] var mutationCount: Int = 0
 
   private var first: List[A] = Nil
   private var last0: ::[A] = null
@@ -50,7 +52,7 @@ class ListBuffer[A]
 
   private type Predecessor[A0] = ::[A0] /*| Null*/
 
-  def iterator = first.iterator
+  def iterator: Iterator[A] = new MutationTracker.CheckedIterator(first.iterator, mutationCount)
 
   override def iterableFactory: SeqFactory[ListBuffer] = ListBuffer
 
@@ -69,7 +71,12 @@ class ListBuffer[A]
     aliased = false
   }
 
-  private def ensureUnaliased() = if (aliased) copyElems()
+  // we only call this before mutating things, so it's
+  // a good place to track mutations for the iterator
+  private def ensureUnaliased(): Unit = {
+    mutationCount += 1
+    if (aliased) copyElems()
+  }
 
   // Avoids copying where possible.
   override def toList: List[A] = {
@@ -97,6 +104,7 @@ class ListBuffer[A]
   }
 
   def clear(): Unit = {
+    mutationCount += 1
     first = Nil
     len = 0
     last0 = null
@@ -301,15 +309,17 @@ class ListBuffer[A]
   }
 
   def mapInPlace(f: A => A): this.type = {
-    ensureUnaliased()
+    mutationCount += 1
     val buf = new ListBuffer[A]
     for (elem <- this) buf += f(elem)
     first = buf.first
     last0 = buf.last0
+    aliased = false // we just assigned from a new instance
     this
   }
 
   def flatMapInPlace(f: A => IterableOnce[A]): this.type = {
+    mutationCount += 1
     var src = first
     var dst: List[A] = null
     last0 = null
@@ -325,6 +335,7 @@ class ListBuffer[A]
       src = src.tail
     }
     first = if(dst eq null) Nil else dst
+    aliased = false // we just rebuilt a fresh, unaliased instance
     this
   }
 
@@ -348,12 +359,24 @@ class ListBuffer[A]
   }
 
   def patchInPlace(from: Int, patch: collection.IterableOnce[A], replaced: Int): this.type = {
-    val i = math.min(math.max(from, 0), length)
-    val n = math.min(math.max(replaced, 0), length)
-    ensureUnaliased()
-    val p = locate(i)
-    removeAfter(p, math.min(n, len - i))
-    insertAfter(p, patch.iterator)
+    val _len = len
+    val _from = math.max(from, 0)         // normalized
+    val _replaced = math.max(replaced, 0) // normalized
+    val it = patch.iterator
+
+    val nonEmptyPatch = it.hasNext
+    val nonEmptyReplace = (_from < _len) && (_replaced > 0)
+
+    // don't want to add a mutation or check aliasing (potentially expensive)
+    // if there's no patching to do
+    if (nonEmptyPatch || nonEmptyReplace) {
+      ensureUnaliased()
+      val i = math.min(_from, _len)
+      val n = math.min(_replaced, _len)
+      val p = locate(i)
+      removeAfter(p, math.min(n, _len - i))
+      insertAfter(p, it)
+    }
     this
   }