--- a/test/files/run/t3088.scala
+++ b/test/files/run/t3088.scala
@@ -2,8 +2,7 @@ import collection.mutable._
 
 object Test {
   def main(args: Array[String]): Unit = {
-    val b = new ListBuffer[Int]
-    b += 1
+    val b = ListBuffer(1, 2, 3)
     b ++= b
   }
 }