Fix appending a `ListBuffer` to itself and subtracting
a `ListBuffer` from itself.

Fix appending a `Growable` that uses the default
`addAll` implementation to itself.

Make `ListBuffer`'s iterator fail-fast when the buffer is
mutated after the iterator's creation.

Fixes scala/bug#3088
Partially addresses scala/bug#12121
This PR is the beginning of a strategy to address scala/bug#12009