@NthPortal 2020-08-18
I'm getting a failure with the following test, and having trouble diagnosing

> scala.tools.nsc.scaladoc.HtmlFactoryTest.(It is not a test it is a sbt.testing.SuiteSelector)

anyone know where to look?