@NthPortal 2020-08-19
benchmarks:

Before
======

```
Benchmark                            (size)  Mode  Cnt         Score        Error  Units
ListBufferBenchmark.filterInPlace        10  avgt   40       209.552 ±      3.828  ns/op
ListBufferBenchmark.filterInPlace       100  avgt   40      1819.224 ±     22.649  ns/op
ListBufferBenchmark.filterInPlace      1000  avgt   40     13115.598 ±   2540.419  ns/op
ListBufferBenchmark.filterInPlace     10000  avgt   40     88563.923 ±   1244.332  ns/op
ListBufferBenchmark.flatMapInPlace1      10  avgt   40       512.107 ±      9.471  ns/op
ListBufferBenchmark.flatMapInPlace1     100  avgt   40      3106.082 ±     38.434  ns/op
ListBufferBenchmark.flatMapInPlace1    1000  avgt   40     40203.206 ±    741.022  ns/op
ListBufferBenchmark.flatMapInPlace1   10000  avgt   40    372594.646 ±  11931.681  ns/op
ListBufferBenchmark.insert               10  avgt   40       233.837 ±      2.661  ns/op
ListBufferBenchmark.insert              100  avgt   40      4695.472 ±     38.096  ns/op
ListBufferBenchmark.insert             1000  avgt   40    476238.273 ±   2797.168  ns/op
ListBufferBenchmark.insert            10000  avgt   40  48167162.543 ± 373388.643  ns/op
ListBufferBenchmark.insertAll            10  avgt   40       313.262 ±      7.922  ns/op
ListBufferBenchmark.insertAll           100  avgt   40      1435.928 ±     20.963  ns/op
ListBufferBenchmark.insertAll          1000  avgt   40     73795.740 ±    626.632  ns/op
ListBufferBenchmark.insertAll         10000  avgt   40   6567366.514 ±  29627.485  ns/op
ListBufferBenchmark.iteratorA            10  avgt   40       215.463 ±      3.137  ns/op
ListBufferBenchmark.iteratorA           100  avgt   40      1805.295 ±     50.502  ns/op
ListBufferBenchmark.iteratorA          1000  avgt   40     17333.637 ±    183.418  ns/op
ListBufferBenchmark.iteratorA         10000  avgt   40     82736.341 ±   1317.157  ns/op
ListBufferBenchmark.iteratorB            10  avgt   40       243.197 ±      4.185  ns/op
ListBufferBenchmark.iteratorB           100  avgt   40      1874.217 ±     32.496  ns/op
ListBufferBenchmark.iteratorB          1000  avgt   40      9206.308 ±    118.528  ns/op
ListBufferBenchmark.iteratorB         10000  avgt   40    142303.132 ±  27946.817  ns/op
ListBufferBenchmark.remove1              10  avgt   40       203.914 ±      3.968  ns/op
ListBufferBenchmark.remove1             100  avgt   40      1327.190 ±     23.052  ns/op
ListBufferBenchmark.remove1            1000  avgt   40    125240.886 ±   1029.013  ns/op
ListBufferBenchmark.remove1           10000  avgt   40  13557831.094 ± 137941.865  ns/op
ListBufferBenchmark.remove2              10  avgt   40       189.867 ±      4.883  ns/op
ListBufferBenchmark.remove2             100  avgt   40       754.743 ±      8.621  ns/op
ListBufferBenchmark.remove2            1000  avgt   40     40683.386 ±    620.152  ns/op
ListBufferBenchmark.remove2           10000  avgt   40   3365176.547 ±  51647.179  ns/op
ListBufferBenchmark.update               10  avgt   40       227.541 ±      1.935  ns/op
ListBufferBenchmark.update              100  avgt   40      4621.739 ±     30.248  ns/op
ListBufferBenchmark.update             1000  avgt   40    480379.841 ±   4256.451  ns/op
ListBufferBenchmark.update            10000  avgt   40  52779922.013 ± 687751.997  ns/op
```

After
=====

```
Benchmark                            (size)  Mode  Cnt         Score        Error  Units
ListBufferBenchmark.filterInPlace        10  avgt   40       196.638 ±      4.577  ns/op
ListBufferBenchmark.filterInPlace       100  avgt   40      1756.866 ±     76.259  ns/op
ListBufferBenchmark.filterInPlace      1000  avgt   40     16077.728 ±    292.931  ns/op
ListBufferBenchmark.filterInPlace     10000  avgt   40     92484.924 ±   1050.074  ns/op
ListBufferBenchmark.flatMapInPlace1      10  avgt   40       485.547 ±      8.411  ns/op
ListBufferBenchmark.flatMapInPlace1     100  avgt   40      3864.339 ±     41.703  ns/op
ListBufferBenchmark.flatMapInPlace1    1000  avgt   40     37358.350 ±    610.488  ns/op
ListBufferBenchmark.flatMapInPlace1   10000  avgt   40    320438.086 ±   4169.805  ns/op
ListBufferBenchmark.insert               10  avgt   40       208.389 ±     10.885  ns/op
ListBufferBenchmark.insert              100  avgt   40      5547.858 ±     70.386  ns/op
ListBufferBenchmark.insert             1000  avgt   40    472445.513 ±   3559.258  ns/op
ListBufferBenchmark.insert            10000  avgt   40  47403816.905 ± 409465.703  ns/op
ListBufferBenchmark.insertAll            10  avgt   40       283.284 ±      6.188  ns/op
ListBufferBenchmark.insertAll           100  avgt   40      2108.278 ±     25.333  ns/op
ListBufferBenchmark.insertAll          1000  avgt   40     73774.687 ±    750.869  ns/op
ListBufferBenchmark.insertAll         10000  avgt   40   6443110.161 ±  57711.887  ns/op
ListBufferBenchmark.iteratorA            10  avgt   40       201.963 ±      9.603  ns/op
ListBufferBenchmark.iteratorA           100  avgt   40      1563.120 ±     28.937  ns/op
ListBufferBenchmark.iteratorA          1000  avgt   40     15574.602 ±    298.189  ns/op
ListBufferBenchmark.iteratorA         10000  avgt   40     86412.182 ±   1132.029  ns/op
ListBufferBenchmark.iteratorB            10  avgt   40       225.307 ±      3.005  ns/op
ListBufferBenchmark.iteratorB           100  avgt   40      1851.259 ±     37.221  ns/op
ListBufferBenchmark.iteratorB          1000  avgt   40     16971.258 ±    285.469  ns/op
ListBufferBenchmark.iteratorB         10000  avgt   40    109750.241 ±   3137.235  ns/op
ListBufferBenchmark.remove1              10  avgt   40       172.079 ±      3.029  ns/op
ListBufferBenchmark.remove1             100  avgt   40      2122.334 ±     23.976  ns/op
ListBufferBenchmark.remove1            1000  avgt   40    127018.261 ±   1097.686  ns/op
ListBufferBenchmark.remove1           10000  avgt   40  13902997.558 ± 100696.734  ns/op
ListBufferBenchmark.remove2              10  avgt   40       164.865 ±      2.265  ns/op
ListBufferBenchmark.remove2             100  avgt   40      1489.532 ±     21.509  ns/op
ListBufferBenchmark.remove2            1000  avgt   40     40996.992 ±    452.266  ns/op
ListBufferBenchmark.remove2           10000  avgt   40   3388047.209 ±  40572.239  ns/op
ListBufferBenchmark.update               10  avgt   40       218.951 ±      8.176  ns/op
ListBufferBenchmark.update              100  avgt   40      5525.563 ±     42.342  ns/op
ListBufferBenchmark.update             1000  avgt   40    490791.195 ±   3660.844  ns/op
ListBufferBenchmark.update            10000  avgt   40  53137869.474 ± 541782.921  ns/op
```