@NthPortal 2020-08-19
```diff
- before
+ after

  Benchmark                         (size)  Mode  Cnt    Score    Error  Units
- ConstructionBenchmark.listBuffer       0  avgt   40   18.419 ±  0.293  ns/op
+ ConstructionBenchmark.listBuffer       0  avgt   40   21.373 ±  0.177  ns/op
- ConstructionBenchmark.listBuffer       1  avgt   40   25.768 ±  0.456  ns/op
+ ConstructionBenchmark.listBuffer       1  avgt   40   28.390 ±  0.273  ns/op
- ConstructionBenchmark.listBuffer      10  avgt   40  114.286 ±  1.931  ns/op
+ ConstructionBenchmark.listBuffer      10  avgt   40  113.392 ±  1.512  ns/op
- ConstructionBenchmark.listBuffer     100  avgt   40  963.965 ± 17.102  ns/op
+ ConstructionBenchmark.listBuffer     100  avgt   40  931.931 ± 17.372  ns/op
```